Feat: 카테고리 추가 모달 새로운 파일로 구분 #19
